fix : 공유 상태 버그 수정 "
