+++ src/1_Page/Profile/ui/TrackingImageTab/index.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import STYLE from "./style";
 import TrackContainer from "./ui/TrackingImageContainer";
 
@@ -6,6 +6,13 @@ const TrackingImageTab = (props) => {
   const { trackingImageList, modifyMode, updateSelectedTracks, obServeRef } =
     props;
 
+  useEffect(() => {
+    console.log("트래킹 탭 렌더링");
+  }, []);
+
+  useEffect(() => {
+    console.log(updateSelectedTracks);
+  }, [updateSelectedTracks]);
   return (
     <>
       {trackingImageList.length === 0 ? (
@@ -15,6 +22,7 @@ const TrackingImageTab = (props) => {
           {trackingImageList.map((trackingImageData, index) => {
             return (
               <TrackContainer
+                key={trackingImageData.idx}
                 obServeRef={
                   index === trackingImageList.length - 1 ? obServeRef : null
                 }
